version_2/pyproject.toml
[project]
name = "initial-version"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.11.0"
dependencies = [
    "google-genai>=1.62.0",
    "pandas>=2.3.3",
    "pydeck>=0.9.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
    "tavily-python>=0.7.21",
]
